vid::avid_m::config
Trait AvidMConfig

pub trait AvidMConfig {
    type BaseField: FftField;
    type Digest: NodeValue;
    type MerkleTree: MerkleTreeScheme<Element = Self::Digest, Commitment = Self::Digest>;

    // Required method
    fn raw_share_digest(
        raw_shares: &[Self::BaseField],
    ) -> Result<Self::Digest, VidError>;
}

Required Associated Types

type BaseField: FftField

type Digest: NodeValue

type MerkleTree: MerkleTreeScheme<Element = Self::Digest, Commitment = Self::Digest>

Required Methods

fn raw_share_digest( raw_shares: &[Self::BaseField], ) -> Result<Self::Digest, VidError>

Digest the raw shares into the element type for Merkle tree.

Errors

This function will return an error if digest function fails.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.