vid::avidm_gf2

Module namespaced

Namespaced AvidmGf2 scheme This file implements the namespaced AvidmGf2 scheme.

Structs

NsAvidmGf2Common

VID Common data that needs to be broadcasted to all storage nodes

NsAvidmGf2Scheme

Dummy struct for namespaced AvidmGf2 scheme

NsAvidmGf2Share

Namespaced share for each storage node, contains one AvidmGf2Share for each namespace.

Type Aliases

NsAvidmGf2Commit

Namespaced commitment type

NsAvidmGf2Param

Namespaced parameter type