vid::avidm_gf2

Module proofs

Namespace proofs for AvidmGf2 scheme This module implements encoding proofs for the Avid-M Scheme.

Structs

NsProof

A proof of a namespace payload. It consists of the index of the namespace, the namespace payload, and a merkle proof of the namespace payload against the namespaced VID commitment.